combine send and receive into a single integration
