+++ src/core_int/integration_slack.py
@@ -5,7 +5,7 @@ import logging
 
 from dataclasses import dataclass
 
-from dazl import create
+from dazl import create, exercise
 from dazl.model.core import ContractData
 from slack import WebClient
 
@@ -18,9 +18,28 @@ from daml_dit_if.main.web import json_response
 LOG = logging.getLogger('integration')
 
 
-def integration_slack_receive_dm_main(
+@dataclass
+class IntegrationSlackSendMessageEnv(IntegrationEnvironment):
+    slackApiToken: str
+
+
+def integration_slack_main(
         env: 'IntegrationEnvironment',
         events: 'IntegrationEvents'):
+
+    sc = WebClient(env.slackApiToken, run_async=True)
+
+    @events.ledger.contract_created(
+        'SlackIntegration.OutboundMessage.OutboundMessage')
+    async def on_contract_created(event):
+        LOG.info('slack send message - created: %r', event)
+
+        channel = event.cdata['slackChannel']
+        text = event.cdata['messageText']
+
+        sc.chat_postMessage(channel=channel, text=text)
+
+        return [exercise(event.cid, 'Archive')]
 
     @events.webhook.post(label='Slack Event Webhook Endpoint')
     async def on_webhook_post(request):
